implement log out function from menu bar, delete btnLogOut from animal group view + controller
