+++ src/ui/view/AnimalGroup.fxml
@@ -24,7 +24,6 @@
         <GridPane alignment="CENTER" hgap="10" layoutX="20.0" layoutY="28.0" prefHeight="550.0" prefWidth="984.0" vgap="10" AnchorPane.bottomAnchor="20.0" AnchorPane.leftAnchor="20.0" AnchorPane.rightAnchor="20.0" AnchorPane.topAnchor="28.0">
 
             <Button fx:id="btnCreate" onAction="#onCreateButtonClicked" prefHeight="26.0" prefWidth="135.0" text="Add" GridPane.halignment="CENTER" GridPane.rowIndex="3" GridPane.valignment="CENTER" />
-            <Button fx:id="btnLogOut" prefHeight="26.0" prefWidth="106.0" text="Log out" GridPane.columnIndex="5" GridPane.valignment="BOTTOM" />
 
             <TableView fx:id="tbAnimalGroup" prefHeight="0.0" prefWidth="982.0" GridPane.columnSpan="6" GridPane.rowIndex="1">
                 <columns>
